plower/requirements.txt
# ~/data/plower/requirements.txt

# FastAPI (APIサーバーフレームワーク)
fastapi
uvicorn

# Gemini API (GoogleのAI SDK)
google-genai

# 環境変数管理
python-dotenv